# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Production-oriented REST API for the OIG Police Management System.",
        "title": "OIG Police Management API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/evidence/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "image/jpeg",
                    "image/png",
                    "image/webp"
                ],
                "tags": [
                    "Penalties"
                ],
                "summary": "Retrieve authorized penalty evidence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Evidence ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/members/by-police-id/{policeId}/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Members"
                ],
                "summary": "Get member disciplinary summary by Police ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Police ID",
                        "name": "policeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.MemberSummary"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/members/{id}/penalties": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penalties"
                ],
                "summary": "List a member's complete penalty history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stable member ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.Penalty"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/penalties": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "actionTypeIds accepts repeated form fields, a JSON string array, or comma-separated IDs/names/legacy PT identifiers. JPEG, PNG, and WebP evidence only.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penalties"
                ],
                "summary": "Create a disciplinary penalty with evidence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Police ID",
                        "name": "policeId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "One or more type identifiers",
                        "name": "actionTypeIds",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Reason",
                        "name": "reason",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "permanent, months, or date_range",
                        "name": "durationType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Months for months duration",
                        "name": "months",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "YYYY-MM-DD; required for date_range",
                        "name": "startDate",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive YYYY-MM-DD; required for date_range",
                        "name": "endDate",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Fine in the smallest configured currency unit",
                        "name": "fineAmount",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "JPEG, PNG, or WebP evidence",
                        "name": "evidence",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreatePenaltyResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/penalties/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penalties"
                ],
                "summary": "Get penalty details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Penalty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.Penalty"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/penalties/{id}/pardon": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penalties"
                ],
                "summary": "Pardon a penalty without deleting its history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Penalty ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pardon reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PardonPenaltyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PardonPenaltyResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/penalty-types": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penalty Types"
                ],
                "summary": "List penalty types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.PenaltyType"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penalty Types"
                ],
                "summary": "Create a penalty type",
                "parameters": [
                    {
                        "description": "Penalty type",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePenaltyTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PenaltyType"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
        "/api/v1/penalty-types/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Penalty Types"
                ],
                "summary": "Delete an unreferenced penalty type",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Penalty type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreatePenaltyResult": {
            "type": "object",
            "properties": {
                "isPermanent": {
                    "type": "boolean"
                },
                "penaltyCount": {
                    "type": "integer"
                },
                "penaltyId": {
                    "type": "string"
                }
            }
        },
        "dto.CreatePenaltyTypeRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "legacyId": {
                    "type": "string",
                    "example": "PT001"
                },
                "name": {
                    "type": "string",
                    "example": "Written warning"
                }
            }
        },
...
        "dto.MemberSummary": {
            "type": "object",
            "properties": {
                "info": {
                    "$ref": "#/definitions/dto.Member"
                },
                "penalties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Penalty"
                    }
                },
                "seriousWarning": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.PardonPenaltyRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "example": "Pardoned after review"
                }
            }
        },
        "dto.PardonPenaltyResult": {
            "type": "object",
            "properties": {
                "penalty": {
                    "$ref": "#/definitions/dto.Penalty"
                },
                "penaltyCount": {
                    "type": "integer"
                }
            }
        },
        "dto.Penalty": {
            "type": "object",
            "properties": {
                "actionTypeIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "actionTypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "durationType": {
                    "type": "string",
                    "enum": [
                        "permanent",
                        "months",
                        "date_range"
                    ]
                },
                "evidenceId": {
                    "type": "string"
                },
                "evidenceUrl": {
                    "type": "string"
                },
                "expiryDate": {
                    "type": "string"
                },
                "fineAmount": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "isPermanent": {
                    "type": "boolean"
                },
                "memberId": {
                    "type": "string"
                },
                "months": {
                    "type": "integer"
                },
                "pardonReason": {
                    "type": "string"
                },
                "pardonedAt": {
                    "type": "string"
                },
                "pardonedBy": {
                    "type": "string"
                },
                "policeId": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                }
            }
        },
...
        "dto.PenaltyType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "legacyId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "message": {
                    "type": "string",
                    "example": "ok"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}